+++ internal/backoff/backoff.go
@@ -1,0 +1,126 @@
+package backoff
+
+import (
+	"math"
+	"time"
+
+	"github.com/ydb-platform/ydb-go-sdk/v3/internal/xrand"
+)
+
+// Backoff is the interface that contains logic of delaying operation retry.
+type Backoff interface {
+	// Wait maps index of the retry to a channel which fulfillment means that
+	// Delay is over.
+	//
+	// Note that retry index begins from 0 and 0-th index means that it is the
+	// first retry attempt after an initial error.
+	Wait(n int) <-chan time.Time
+
+	// Delay returns mapping of i to Delay.
+	Delay(i int) time.Duration
+}
+
+// Default parameters used by Retry() functions within different sub packages.
+const (
+	fastSlot = 5 * time.Millisecond
+	slowSlot = 1 * time.Second
+)
+
+var (
+	Fast = New(
+		WithSlotDuration(fastSlot),
+		WithCeiling(6),
+	)
+	Slow = New(
+		WithSlotDuration(slowSlot),
+		WithCeiling(6),
+	)
+)
+
+// logBackoff contains logarithmic Backoff policy.
+type logBackoff struct {
+	// SlotDuration is a size of a single time slot used in Backoff Delay
+	// calculation.
+	// If SlotDuration is less or equal to zero, then the time.Second value is
+	// used.
+	SlotDuration time.Duration
+
+	// Ceiling is a maximum degree of Backoff Delay growth.
+	// If Ceiling is less or equal to zero, then the default ceiling of 1 is
+	// used.
+	Ceiling uint
+
+	// JitterLimit controls fixed and random portions of Backoff Delay.
+	// Its value can be in range [0, 1].
+	// If JitterLimit is non zero, then the Backoff Delay will be equal to (F + R),
+	// where F is a result of multiplication of this value and calculated Delay
+	// duration D; and R is a random sized part from [0,(D - F)].
+	JitterLimit float64
+
+	// generator of jitter
+	r xrand.Rand
+}
+
+type option func(b *logBackoff)
+
+func WithSlotDuration(slotDuration time.Duration) option {
+	return func(b *logBackoff) {
+		b.SlotDuration = slotDuration
+	}
+}
+
+func WithCeiling(ceiling uint) option {
+	return func(b *logBackoff) {
+		b.Ceiling = ceiling
+	}
+}
+
+func WithJitterLimit(jitterLimit float64) option {
+	return func(b *logBackoff) {
+		b.JitterLimit = jitterLimit
+	}
+}
+
+func New(opts ...option) logBackoff {
+	b := logBackoff{
+		r: xrand.New(xrand.WithLock()),
+	}
+	for _, o := range opts {
+		o(&b)
+	}
+	return b
+}
+
+// Wait implements Backoff interface.
+func (b logBackoff) Wait(n int) <-chan time.Time {
+	return time.After(b.Delay(n))
+}
+
+// Delay returns mapping of i to Delay.
+func (b logBackoff) Delay(i int) time.Duration {
+	s := b.SlotDuration
+	if s <= 0 {
+		s = time.Second
+	}
+	n := 1 << min(uint(i), max(1, b.Ceiling))
+	d := s * time.Duration(n)
+	f := time.Duration(math.Min(1, math.Abs(b.JitterLimit)) * float64(d))
+	if f == d {
+		return f
+	}
+	return f + time.Duration(b.r.Int64(int64(d-f)+1))
+}
+
+func min(a, b uint) uint {
+	if a < b {
+		return a
+	}
+	return b
+}
+
+func max(a, b uint) uint {
+	if a > b {
+		return a
+	}
+	return b
+}