+++ internal/table/retry.go
@@ -4,6 +4,7 @@ import (
 	"context"
 	"fmt"
 
+	"github.com/ydb-platform/ydb-go-sdk/v3/internal/backoff"
 	"github.com/ydb-platform/ydb-go-sdk/v3/internal/xerrors"
 	"github.com/ydb-platform/ydb-go-sdk/v3/retry"
 	"github.com/ydb-platform/ydb-go-sdk/v3/table"
@@ -217,29 +218,22 @@ func (s *singleSession) CloseSession(ctx context.Context, x Session) error {
 func retryBackoff(
 	ctx context.Context,
 	p SessionProvider,
-	fastBackoff retry.Backoff,
-	slowBackoff retry.Backoff,
+	fastBackoff backoff.Backoff,
+	slowBackoff backoff.Backoff,
 	isOperationIdempotent bool,
 	op table.Operation,
 ) (err error) {
 	var (
-		s        Session
-		i        int
-		attempts int
-		code     = int64(0)
+		s Session
 	)
 	defer func() {
 		if s != nil {
 			_ = p.Put(ctx, s)
 		}
 	}()
-	for ; ; i++ {
-		attempts++
-		select {
-		case <-ctx.Done():
-			return xerrors.WithStackTrace(ctx.Err())
-
-		default:
+	return retry.Retry(
+		ctx,
+		func(ctx context.Context) (err error) {
 			if s == nil {
 				s, err = p.Get(ctx)
 				if s == nil && err == nil {
@@ -263,24 +257,15 @@ func retryBackoff(
 
 			m := retry.Check(err)
 
-			if m.StatusCode() != code {
-				i = 0
-			}
-
 			if m.MustDeleteSession() && s != nil {
 				_ = p.CloseSession(ctx, s)
 				s = nil
 			}
 
-			if !m.MustRetry(isOperationIdempotent) {
-				return xerrors.WithStackTrace(err)
-			}
-
-			if retry.Wait(ctx, fastBackoff, slowBackoff, m, i) != nil {
-				return xerrors.WithStackTrace(err)
-			}
-
-			code = m.StatusCode()
-		}
-	}
+			return xerrors.WithStackTrace(err)
+		},
+		retry.WithInternalFastBackoff(fastBackoff),
+		retry.WithInternalSlowBackoff(slowBackoff),
+		retry.WithIdempotent(isOperationIdempotent),
+	)
 }