+++ retry/retry.go
@@ -2,30 +2,25 @@ package retry
 
 import (
 	"context"
-	"math"
 	"time"
 
+	"github.com/ydb-platform/ydb-go-sdk/v3/internal/backoff"
+	"github.com/ydb-platform/ydb-go-sdk/v3/internal/retry"
+	"github.com/ydb-platform/ydb-go-sdk/v3/internal/wait"
 	"github.com/ydb-platform/ydb-go-sdk/v3/internal/xerrors"
-	"github.com/ydb-platform/ydb-go-sdk/v3/internal/xrand"
 	"github.com/ydb-platform/ydb-go-sdk/v3/trace"
 )
 
-// Default parameters used by Retry() functions within different sub packages.
-const (
-	fastSlot = 5 * time.Millisecond
-	slowSlot = 1 * time.Second
-)
-
-// Default parameters used by Retry() functions within different sub packages.
 var (
-	FastBackoff = newBackoff(
-		withSlotDuration(fastSlot),
-		withCeiling(6),
-	)
-	SlowBackoff = newBackoff(
-		withSlotDuration(slowSlot),
-		withCeiling(6),
-	)
+	// FastBackoff is a default fast backoff object
+	// Deprecated: don't use explicit it, will be removed at next major release.
+	// Use retry.WithFastBackoff retry option for redefine default fast backoff instead
+	FastBackoff = backoff.Fast
+
+	// SlowBackoff is a default fast backoff object
+	// Deprecated: don't use explicit it, will be removed at next major release.
+	// Use retry.WithSlowBackoff retry option for redefine default slow backoff instead
+	SlowBackoff = backoff.Slow
 )
 
 // retryOperation is the interface that holds an operation for retry.
@@ -36,21 +31,25 @@ type retryOperation func(context.Context) (err error)
 type retryableErrorOption xerrors.RetryableErrorOption
 
 const (
-	BackoffTypeNoBackoff   = xerrors.BackoffTypeNoBackoff
-	BackoffTypeFastBackoff = xerrors.BackoffTypeFastBackoff
-	BackoffTypeSlowBackoff = xerrors.BackoffTypeSlowBackoff
+	BackoffTypeNoBackoff   = backoff.TypeNoBackoff
+	BackoffTypeFastBackoff = backoff.TypeFastBackoff
+	BackoffTypeSlowBackoff = backoff.TypeSlowBackoff
 )
 
-func WithBackoff(t xerrors.BackoffType) retryableErrorOption {
+// WithBackoff makes retryable error option with custom backoff type
+func WithBackoff(t backoff.Type) retryableErrorOption {
 	return retryableErrorOption(xerrors.WithBackoff(t))
 }
 
+// WithDeleteSession makes retryable error option with delete session flag
 func WithDeleteSession() retryableErrorOption {
 	return retryableErrorOption(xerrors.WithDeleteSession())
 }
 
+// RetryableError makes retryable error from options
+// RetryableError provides retrying on custom errors
 func RetryableError(err error, opts ...retryableErrorOption) error {
-	return xerrors.RetryableError(
+	return xerrors.Retryable(
 		err,
 		func() (retryableErrorOptions []xerrors.RetryableErrorOption) {
 			for _, o := range opts {
@@ -65,8 +64,8 @@ type retryOptions struct {
 	id          string
 	trace       trace.Retry
 	idempotent  bool
-	fastBackoff Backoff
-	slowBackoff Backoff
+	fastBackoff backoff.Backoff
+	slowBackoff backoff.Backoff
 
 	panicCallback func(e interface{})
 }
@@ -88,21 +87,43 @@ func WithTrace(trace trace.Retry) retryOption {
 }
 
 // WithIdempotent returns idempotent option
-func WithIdempotent() retryOption {
+func WithIdempotent(idempotent bool) retryOption {
 	return func(h *retryOptions) {
-		h.idempotent = true
+		h.idempotent = idempotent
 	}
 }
 
-// WithFastBackoff returns fast backoff option
-func WithFastBackoff(b Backoff) retryOption {
+// WithFastBackoff replaces default fast backoff
+func WithFastBackoff(slotDuration time.Duration, ceiling uint, jitterLimit float64) retryOption {
+	return func(h *retryOptions) {
+		h.fastBackoff = backoff.New(
+			backoff.WithSlotDuration(slotDuration),
+			backoff.WithCeiling(ceiling),
+			backoff.WithJitterLimit(jitterLimit),
+		)
+	}
+}
+
+// WithSlowBackoff replaces default slow backoff
+func WithSlowBackoff(slotDuration time.Duration, ceiling uint, jitterLimit float64) retryOption {
+	return func(h *retryOptions) {
+		h.slowBackoff = backoff.New(
+			backoff.WithSlotDuration(slotDuration),
+			backoff.WithCeiling(ceiling),
+			backoff.WithJitterLimit(jitterLimit),
+		)
+	}
+}
+
+// WithInternalFastBackoff replaces default fast backoff
+func WithInternalFastBackoff(b backoff.Backoff) retryOption {
 	return func(h *retryOptions) {
 		h.fastBackoff = b
 	}
 }
 
-// WithSlowBackoff returns slow backoff option
-func WithSlowBackoff(b Backoff) retryOption {
+// WithInternalSlowBackoff replaces default slow backoff
+func WithInternalSlowBackoff(b backoff.Backoff) retryOption {
 	return func(h *retryOptions) {
 		h.slowBackoff = b
 	}
@@ -124,8 +145,8 @@ func WithPanicCallback(panicCallback func(e interface{})) retryOption {
 // If you need to retry your op func on some logic errors - you must return RetryableError() from retryOperation
 func Retry(ctx context.Context, op retryOperation, opts ...retryOption) (err error) {
 	options := &retryOptions{
-		fastBackoff: FastBackoff,
-		slowBackoff: SlowBackoff,
+		fastBackoff: backoff.Fast,
+		slowBackoff: backoff.Slow,
 	}
 	for _, o := range opts {
 		o(options)
@@ -173,7 +194,7 @@ func Retry(ctx context.Context, op retryOperation, opts ...retryOption) (err err
 				return xerrors.WithStackTrace(err)
 			}
 
-			if e := Wait(ctx, options.fastBackoff, options.slowBackoff, m, i); e != nil {
+			if e := wait.Wait(ctx, options.fastBackoff, options.slowBackoff, m.BackoffType(), i); e != nil {
 				return xerrors.WithStackTrace(err)
 			}
 
@@ -185,171 +206,12 @@ func Retry(ctx context.Context, op retryOperation, opts ...retryOption) (err err
 }
 
 // Check returns retry mode for err.
-func Check(err error) (m retryMode) {
-	statusCode, operationStatus, backoff, deleteSession := xerrors.Check(err)
-	return retryMode{
-		statusCode:      statusCode,
-		operationStatus: operationStatus,
-		backoff:         backoff,
-		deleteSession:   deleteSession,
-	}
-}
-
-func Wait(ctx context.Context, fastBackoff Backoff, slowBackoff Backoff, m retryMode, i int) error {
-	var b Backoff
-	switch m.BackoffType() {
-	case xerrors.BackoffTypeNoBackoff:
-		return nil
-	case xerrors.BackoffTypeFastBackoff:
-		b = fastBackoff
-	case xerrors.BackoffTypeSlowBackoff:
-		b = slowBackoff
-	}
-	return waitBackoff(ctx, b, i)
-}
-
-// logBackoff contains logarithmic Backoff policy.
-type logBackoff struct {
-	// SlotDuration is a size of a single time slot used in Backoff delay
-	// calculation.
-	// If SlotDuration is less or equal to zero, then the time.Second value is
-	// used.
-	SlotDuration time.Duration
-
-	// Ceiling is a maximum degree of Backoff delay growth.
-	// If Ceiling is less or equal to zero, then the default ceiling of 1 is
-	// used.
-	Ceiling uint
-
-	// JitterLimit controls fixed and random portions of Backoff delay.
-	// Its value can be in range [0, 1].
-	// If JitterLimit is non zero, then the Backoff delay will be equal to (F + R),
-	// where F is a result of multiplication of this value and calculated delay
-	// duration D; and R is a random sized part from [0,(D - F)].
-	JitterLimit float64
-
-	// generator of jitter
-	r xrand.Rand
-}
-
-type option func(b *logBackoff)
-
-func withSlotDuration(slotDuration time.Duration) option {
-	return func(b *logBackoff) {
-		b.SlotDuration = slotDuration
-	}
-}
-
-func withCeiling(ceiling uint) option {
-	return func(b *logBackoff) {
-		b.Ceiling = ceiling
-	}
-}
-
-func withJitterLimit(jitterLimit float64) option {
-	return func(b *logBackoff) {
-		b.JitterLimit = jitterLimit
-	}
-}
-
-func newBackoff(opts ...option) logBackoff {
-	b := logBackoff{
-		r: xrand.New(xrand.WithLock()),
-	}
-	for _, o := range opts {
-		o(&b)
-	}
-	return b
-}
-
-// Wait implements Backoff interface.
-func (b logBackoff) Wait(n int) <-chan time.Time {
-	return time.After(b.delay(n))
-}
-
-// delay returns mapping of i to delay.
-func (b logBackoff) delay(i int) time.Duration {
-	s := b.SlotDuration
-	if s <= 0 {
-		s = time.Second
-	}
-	n := 1 << min(uint(i), max(1, b.Ceiling))
-	d := s * time.Duration(n)
-	f := time.Duration(math.Min(1, math.Abs(b.JitterLimit)) * float64(d))
-	if f == d {
-		return f
-	}
-	return f + time.Duration(b.r.Int64(int64(d-f)+1))
-}
-
-func min(a, b uint) uint {
-	if a < b {
-		return a
-	}
-	return b
-}
-
-func max(a, b uint) uint {
-	if a > b {
-		return a
-	}
-	return b
-}
-
-// retryMode reports whether operation is able retried and with which properties.
-type retryMode struct {
-	statusCode      int64
-	operationStatus xerrors.OperationStatus
-	backoff         xerrors.BackoffType
-	deleteSession   bool
-}
-
-func (m retryMode) MustRetry(isOperationIdempotent bool) bool {
-	switch m.operationStatus {
-	case xerrors.OperationFinished:
-		return false
-	case xerrors.OperationStatusUndefined:
-		return isOperationIdempotent
-	default:
-		return true
-	}
-}
-
-func (m retryMode) StatusCode() int64 { return m.statusCode }
-
-func (m retryMode) MustBackoff() bool { return m.backoff&xerrors.BackoffTypeBackoffAny != 0 }
-
-func (m retryMode) BackoffType() xerrors.BackoffType { return m.backoff }
-
-func (m retryMode) MustDeleteSession() bool { return m.deleteSession }
-
-// Backoff is the interface that contains logic of delaying operation retry.
-type Backoff interface {
-	// Wait maps index of the retry to a channel which fulfillment means that
-	// delay is over.
-	//
-	// Note that retry index begins from 0 and 0-th index means that it is the
-	// first retry attempt after an initial error.
-	Wait(n int) <-chan time.Time
-}
-
-// waitBackoff is a helper function that waits for i-th Backoff b or ctx
-// expiration.
-// It returns non-nil error if and only if deadline expiration branch wins.
-func waitBackoff(ctx context.Context, b Backoff, i int) error {
-	if b == nil {
-		if err := ctx.Err(); err != nil {
-			return xerrors.WithStackTrace(err)
-		}
-		return nil
-	}
-	select {
-	case <-b.Wait(i):
-		return nil
-	case <-ctx.Done():
-		if err := ctx.Err(); err != nil {
-			return xerrors.WithStackTrace(err)
-		}
-		return nil
-	}
+func Check(err error) (m retry.Mode) {
+	statusCode, operationStatus, backoff, deleteSession := retry.Check(err)
+	return retry.NewMode(
+		statusCode,
+		operationStatus,
+		backoff,
+		deleteSession,
+	)
 }